v4.3.1c BUG FIX:
Few updates to the code.
- ED_BATH_USER.f90: the default symmetries operations have been extended to the
lattice case, i.e. deal with Nineq bath arrays. This fix a bug related to the
writing of the bath to a file. This operation was left to the user, but in the
case of Nineq inequivalent sites each bath must be wrote to a different file
labelled with the ed_suffix=_siteXXXX .
If (like me AA) the user forget to do this after a symmetry
operation is applied (indeed the case for me, like spin symmetrization)
the system ends up writing all the symmetrized bath to the same file, corresponding to
the last value of the ed_suffix variable (different for each process and equal to the
last site worked out by the thread). This could lead to problems in the restart procedure.
Now the set of ineq. baths are handled by the code directly.

-ED_CHI2FIT: ed_suffix is reset to empty string after fit is done.

- ed_bhz_2d_edge: now has a default Left-Right symmetry and works out only the
Inequivalent Ly/2 sites.