v4.3.2 DEBUG!!
AV found several bugs in the updated version of the code.
AA & AV worked together to fix all the bugs.

AV noted that his driver for the nano-flake in the SC channel did NOT work.
A first analysis pointed out that the impSigma functions were NaN!!!
To solve the problem we restarted from the repulsive/attractive code
for the homogeneous square lattice and fixed one by one all the bugs:

- in ED_SETUP.f90 the procedures setup_pointers_{normal,superc,nonsu2} had a wrong
statement:
>
>     do isector=1,Nsectors
>-         neigen_sector(isector) = min(dim,lanc_nstates_sector)   !init every sector to required eigenstates
>+         neigen_sector(isector) = min(getdim(isector),lanc_nstates_sector)   !init every sector to required eigenstates
>     enddo
the variable dim took its last value in a loop appearing above this statement, causing lanc_nstates_sector to be equal to it,
i.e. identically 1 for all sectors. Although this did not screw up the calculations it definitively slowed them down,
as the lanc_nstates_sector had to be increased starting from 1 rather than its default value 6.
In addition this forced the system to include in the state_list states from sectors very far in energy from the groundstate,
causing the Boltzmann factor exp(-beta*(Ei-Egs)) to reach its underflow value (< tiny(1d0)).
This was recognized by FP who however, rather than identify and solve the original problem introduced a series of
if conditions. Too bad, now it is fixed :-) :-)

- in ED_GREENS_FUNCTIONS.f90/ed_greens_funcs_build_gf_superc.f90 this FUNDAMENTAL line has been deleted (WTF!!!)
>    integer                                    :: i,j,ierr
>    complex(8)                                 :: iw
>    !
> +  Egs = state_list%emin       !get the gs energy
>    !
>    if((finiteT).and.(beta*(Ei-Egs).lt.200))then
>      pesoBZ = vnorm2*exp(-beta*(Ei-Egs))/zeta_function
>    elseif(.not.finiteT)then
>      pesoBZ = vnorm2/zeta_function
>    else
>      pesoBZ=0.d0
>    endif
so Egs was not defined, causing the third option to be true in some cases, which indeed is clearly very very bad.
Indeed pesoBZ should NEVER be equated to zero! But for the time being we left this statement as it is, because if
we do things carefully enough only the first two conditions are met.

The tests for repulsive/attractive Hubbard square lattice (ed_hm_square, ed_ahm_square) gave very good results.
Tests are in order for the ed_nano code (AV).

TODO:
We noted also a bug in the ED_CHI2FIT module, where the Weiss/Delta function is defined with dimensions different from
those actually passed by reference in the call. This needs to be fixed as soon as possible!

AA - AV