Minor changes to ED_SETUP.f90 that dd not coplile with gfortran due to improper format of info string.
Minor changes to the drivers:
 - ed_ahm_square.f90:
   commented out the get_kin functions and relative calls
   that prevented the code to compile
 - ed_haldane.f90
   this also ha the bug affetting the ed_graphene.f90 (see below)

Added a new driver ed_graphene.f90 which is based on ed_haldane.f90
but can also treat the symmetry-broken Nspin=2 case.
There is still a bug in the sampling of the Brillouin zone
for lattices other than cubic, resulting in non-zero elements of Hloc
and wiggles in the local non-interacting DOS of the model.
The issue should be solved by generating the proper lattice kgrids
using the reciprocal lattice vectors rather than using the cubic mesh.
The efforts so far managed to reduce the error, but not eliminating it.

In order to have a clean graphene code one should either purge the
Haldane variables (tsp, mh, phi) and related input files
or merge the code back to ed_haldane.f90 and have a unique driver.