1st Commit in the dev_MPI_directHxV_LessMemory branch.
This development line is for:
a. put the code on diet, using the global interface/wrapper to SPARSE LINear ALGebra
now included in SciFortran (SF_SP_LINALG.f90). We included there all the interfaces
to P-Arpack and Plain Lanczos. This includes a better handling of the Mat-Vec products.

b. setup the code so to perform direct HxV products: Do Not Store the Hamiltonian anymore.
This will serve as a basis to do larger systems calculations.

c. pass the code to the nup-ndw basis in the setup of the sector. For more info
read the README.md file.

> now:
The SF_SP_LINALG.f90 has been extensively tested in the serial/parallel channels
(see officina/ARPACK_DETACH). Perfect resutls so far, including better handling
of the matrix-vector product interface in the parallel case, which are
now based on the MPI_AllGatherV call (less memory usage).
The code compile with updted ED_MATVEC.f90 having all the calls to arpack and lanczos
been replaced with the corresponding functions in SciFor (sp_eigh, sp_lanc_tridiag).