Pause here:
- the code has been completely updated with a single universal MPI implementation.
- the code uses external, more efficient and more compact P-Arpack and Lanczos implementation (stored in SciFortran).
This features have been extensively tested.
- some parts of the code are ready to be deleted, such as ED_GLOC, ED_ENERGY and ED_WEISS, which are now integrated and generalized
in DMFT_TOOLS (final tests to be done, so wait).

After testing the code is ready to be updated with direct HxV matrix-vector multiplication AND nup-ndw basis redefinition.
This work will enable a ridiculously small use of the memory: no store of the sparse H, no store of the direct Sector-Hilbert space map (Hmap).
The latter will be substituted with a direct and inverse on-the-fly look-up, based on the reduced basis nup-ndw.
The code will be slightly slower but will easily access larger systems (Ns=15-18). Speed can be improved with solid parallelization.

Ciao. Now I have to work.