5th intermediate step. Nup-Ndw factorization implemented in the ed_mode=normal and bath_type=irred/hybrid channels.
The updated is however **not complete**, because it still relies
on a full sparse Hamiltonian of the sector. The latter can be
factorized in two sub-matrices (one per spin type), which however
requires to act on the mat-vec product as well &rarr; this is
post-poned to the next test (test 6).

In order to achieve a better handling of the Hamiltonian,
I scattered the 2 available procedure build_H_d,c into different
cases, according to the value of the variable ed_mode and
 bath_type. So now we have multiple procedures as follows:

* ed_mode=**normal** ; bath_type=**irred/hybrd**. [*updated to N^up-N^dw* ]
* ed_mode=**normal** ; bath_type=**replica**      [*not yet updated*]
* ed_mode=**superc** ; bath_type=**irred/hybrd**  [*as before*]
* ed_mode=**superc** ; bath_type=**replica**      [*as before*]
* ed_mode=**nonsu2** ; bath_type=**irred/hybrd**  [*as before*]
* ed_mode=**nonsu2** ; bath_type=**replica**      [*as before*]

**RESULTS**
After some simple debug results of the run in test_5 gives identical
data as test_4.2. Difference are at machine precision (last digit
in the impurity functions). Numerically tolerant in eps=1.d-14
comparision returns identical results. eps must be increased for
some, more sensible data, such as hamiltonian.restart.

In the next step we shall:
- complete the implementation in the normal-irrhyb channels
- extend the implementation to the replica case.
- clean the code a bit.