6th. this is an intermediate step in the factorized Nup-Ndw algorithm. Hamiltonian building and Mat-Vec product have been adapted. Code compiles. The implementation is not straightforward and requires further development to be really optimized. Especially the actual implementation had to sacrifice MPI parallelization.
In order to study better the situation and avoid huge
code regression I shall *rebase* this branch, i.e.
I shall move the changes up to the last or last two
committment (see log) into a separate branch and keep the last or
last two for further development of the NupNdw only.