v4.4.1 BUG FIX & UPDATE
We (AA & AV) fixed a bug introduced in the recent NONSU2 development.
The bug concerned a wrong construction of the Hamiltonian H in the
normal channel.

We cleaned the code structure of the build_H_normal.f90 file, included
in ED_HAMILTONIAN, which creates the H for any given sector.

The erroneous lines of code enable the possibilty to have
spin-flip hopping even in the normal channel (where they are
forbidden by symmetry). This bug was raised by a non perfect
cancellation of the local impurity Hamiltonian for the AFM
graphene model (which has Nspin=2 and a finite residual value
of impHloc(sigma,sigma',alfa,beta)).

The code (as written by fpetox) is definitively correct
for the Nspin=2 case by DOES NOT WORK for the NSPIN=1, because
he is not taking into account the fact that the structure of the
states alwyas has spins UP and DW.
Hopping events for both spins must be taken into account, irrespective
of the value of NSPIN. So the usual loop
do ispin=1,Nspin
...
enddo
can not be used! (it works only for Nspin=2)

With the occasion we simplified the code structure of ED_HAMILTONIAN a bit
by including different build_H_xxx.f90 for each value of ed_mode=normal,superc,nonsu2.
Procedure pointers have been set up in ED_MAIN.

CIAO