v4.4.2 UPDATE (forced):
** Forced update:
Added some features to BHZ-EDGE-AFM code: to evaluate
the renormalized topological Hamiltonian H_top = Z*[Hk + ReSigma(0)]

* Forced update:
After many tests the driver for the BHZ-EDGE-AFM case is working.
In the process I introduced also a driver to solve the homogenenuous case
(PBC in all direction) for the 2-atoms in the unit cell case, which compare exactly
with the former 4-atoms driver.

Introduced a new driver for BHZ 2D in the STRIPE geometry and AFM ordering:
ed_bhz_2d_edge_afm2 (2 stands for 2 atoms in the unit cell).

Some BHZ drivers have been renamed.

In the process, some internal routines and variables have been made completely
opaque to the user.
In particular *Nlat* is now removed from ED_VARS_GLOBAL as the number of
inequivalent sites to be solved is decided case by case in each routine.

Initial tests for BHZ_2D_EDGE_AFM2 give excellent results which numerically
agree with former results for the stripe.