Commit before pulling the master branch.
Corrected a bug in the ed_haldane and ed_haldance_sc drivers
in the definition of the topological Hamiltonian, leading to
a change of curvature in the transition line between the two
band insulator and Chern insulator phases.

The correct definition is:
Eigvec = Hk(:,:,ik) + nnn2lso_reshape(S0,Nlat,Nspin,Norb)

Also, restored the positive sign of the phase \phi in the
Haldane Hamiltonian, as the wrong value of the Chern number
was probably due to the incorrect sign in the calculation
of the Barry phase, as the path in the plaquette integration
is implemented clockwise rather than anticlockwise.