UPDATE - BUG FOUND (SCIFOR) - NOT FIXED
FP found a bug, most probably located in ED_MATVEC and supposedly trackable to
the MPI_allgatherv  call.
This procedure (MPI) seems NOT to work in OPEN-MPI implementation (tested on
1.8.3[@Ulysses-HPC] and 1.8.8[@JARVIS]).
As it went unnoticed in MPICH implementation of MPI we deduce that there is
something wrong with OPENMPI only. PLEASE USE MPICH is running this code.

The error is a mysterious segmentation fault, signal 11.
The error has been reported using ed_SOC driver and it appears also using
one CPU (not mpiexec of the code).