Fixed format incompatible with gfortran: stop" > stop " occurring in several ED_ files.
Updated the ed_nano and ed_nano_isoc drivers, as follows:

Defined independent read_sigma and save_sigma functions for both
matsubara and real self-energy.

Implemented the (optional, controlled by the flag "para")
spin- symmetrization of the inequivalent self-energy, to enforce
the paramagnetic solution in the SOC driver.
Setting the flag true works for Nspin=2, but raises an error otherwise.
Note: as the driver is still fitting the spin components separately,
itrasults in a numerical spin-imbalance in the self-energy, which
is evidennt in the spin-resolved conductance (in the log scale plot).

Fixed bug in the usage of the leads' hybridization functions.
In the drivers, both for a matsubara and real frequency objects
were created; however the routine dmft_set_gamma can set only
one hybridization function at a time.
In the driver dmft_set_gamma ow is called every time, to set
the required hybridization in matsubara or real frequency
(e.g., to get the Green's function or transmission coefficient).

Found a (series of) bug(s) in the setting of the hybridization function
for the calculations of the transmission function.
1) broad-band limit DOS: a factor pi is needed to ensure normalization;
2) removed hard-coded symmetrization i<>j and iorb<>jorb in the
   hybridization function, now the symmetry must be given in the input;
3) changed dimension of lmask and rmask from (:) to (:,:)
   now they both carry a pair of indices (i,j) to identify all possible
   hybridization processes, bot local (i=j) and non-local (i/=j) ones;
   also here all symmetries must be given in the inuput;

Note: All those bugs affected the broadening of the Green's funtion,
and the normalization of the transmission coefficent.
However, none of them change the physics of the QI antiresonance
(for very subtle reasons, dictated by the chiral sublattice symmetry).
Hence, the calculations done so far in the broad-band limit are fine.
It could have been a problem for a different hybridization DOS.