Included another driver ed_nano_fix_trans in which the bugs documented in the previous commit were solved.
The tests showed that the quantum interference features found
in the transmission coefficient of graphene nanoflakes are
preserved both in the paramagnetic and antiferromagnetic state,
apart from a meaningless change in the overall value of the
transmission coefficient.

Soon the fix will be transferred to the ed_nano and ed_nano_isoc
stable version of the drivers.