UPDATE: New parallel scheme for GF evaluation.
The Lanczos scheme for the GF and other local dynamical functions calculations
has been updated. The new scheme has a better parallel strategy, identical to
that of ARPACK, in which Lanczos steps are performed on reduced vectors (Nloc<<N).

The new Lanczos (ARPACK and Plain Lanczos) schemes are imported in SciFor in
SF_SP_LINALG and have been thoroughly tested in ARPACK_DETACH @ officina.

The update mainly involved the following steps:
- reducing the number of MatVec routines removing the lanc_xyz ones in favor
of only one type (dd,dc,cc) used both in Arpack and Lanczos.
- Updating all the codes in ED_GREEENS_FUNCTIONS to cope with the new parallel
strategy and new routines.

The rest of the code is essentially untouched. I commented the Weights&Poles
part of the code which look rather useless so far.