UPDATE:
Tests of this new version devel_MPI_GF (see previous commit)
performed using:
- ed_hm_bethe VS ed_hm_bethe_mpi
- ed_bhz_2d (MPI)

Results are in full agreement with previous ones.
There is a small mismatch in the local functions impG --> impSigma
of order O(10D-12). This lead to a small mismatch in the updated Bath.

We could not trace back the origin of this negligible mismatch, so
we assume this is due to the "improved" Lanczos :-)

In the debug process we found a small bug in ED_MAIN:
the call to observables_impurity is before that to build_gf_impurity
leading to an errore in the evaluation of Z and ScatteringRate which
uses the  self-energy of the previous loop (or zero!).
Fixed by shifting down observables_impurity.

This branch will be merged into devel, i.e. the development branch
split from the master.
In *devel we shall import all the updates scheduled for this time.