UPDATE:
Some updates to the lastest version of the code
- Introduced a modification to the interaction part of the Hamiltonian.
The S-E and P-E coupling costants, Jx and Jp resep., are now independent from Jhund as in a
generalized version of the Kanamori Hamiltonian.
NOTE THAT U(1)xSU(2)xSO(2,3) IS OBTAINED ONLY FOR Jx=Jp=Jhund as explained in the paper
"Strong correlations from Hund’s coupling" Georges, de'Medici & Mvralje (ann. of phys.)

This comes handy when you have to tweak the interaction term for some problems.

- fixed some bugs in the ED_OBSERVABLES and build_gf_superc where vectors vvinits
used for application of the operators in the evaluation, resp., phi and GF, were
not initialized properly to cmplx(0d0,0d0) but only to dble(0d0). This led to
bad computation of phi in the two orbital case for example.

- other minor issues.