update, fixed regression and bug in Hbath/HxVbath
- in the reduction process there an error was introduced (by me, AA) in the
evaluation of the bath term of the Hamiltonian for the bath_type=replica.
The code was not working as reported by F.Petocchi.
Now this should be fixed (waiting testing from FP) by simply copying and
adapting the former part of the code from build_h_all.f90 (known to work).