v5.0.0 NEW VERSION
amend 1: fixed compilation on Intel platform.
amend 2: updated Makefile

This is possibly the final commit of the updated version of the ED code.
This code compiles and run using the latest versions of Scifor & Dmft_Tools.

Installation notes:

1. installation via cmake is now possible. This method automatically check
dependencies on the system (scifor and dmft_tools).
 a. <in source>
$ cmake . [-DEXE=ed_hm_bethe/put-driver-here] [-DUSE_MPI=ON/off] [-DBUILD_TYPE=TESTING/debug/release] [-DVERBOSE=ON/off] [-DPREFIX=.bin/path-to-dir]
$ make
 b. <out source>
$ mkdir build
$ cd build
$ cmake .. [-DEXE=ed_hm_bethe/put-driver-here] [-DUSE_MPI=ON/off] [-DBUILD_TYPE=TESTING/debug/release] [-DVERBOSE=ON/off] [-DPREFIX=.bin/path-to-dir]
$ make

2. installation via the usual Makfile is now achieved using the Makefile.inc file.
Renaming of the makefile is necessary to avoid conflict with Cmake
$ make -f Makefile.inc