v5.1.1 UPDATE
In order to avoid messed up printing on the std output from all
threads in a parallel MPI run a different setup is now used.

Each thread is assigned with a unit where to print as a standard output,
this is conventionally named *LOGfile*. The master has LOGfile=6=stdout.
The other have a decreasing number associated to opened files named:
stdOUT_rank<rank of the thread>.
In this way one can check what is currently a thread doing or reading
what did afterward.

Some fine tuning is expected in the near future.